refactor(java): pull common Java release PR logic into shared base class
