+++ src/releasers/java-yoshi.ts
@@ -12,13 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {ReleasePR, ReleaseCandidate} from '../release-pr';
+import {ReleaseCandidate, PackageName} from '../release-pr';
 
-import {ConventionalCommits} from '../conventional-commits';
-import {GitHubTag} from '../github';
-import {checkpoint, CheckpointType} from '../util/checkpoint';
 import {Update, VersionsMap} from '../updaters/update';
-import {Commit} from '../graphql-to-commits';
 
 // Generic
 import {Changelog} from '../updaters/changelog';
@@ -27,141 +23,19 @@ import {GoogleUtils} from '../updaters/java/google-utils';
 import {PomXML} from '../updaters/java/pom-xml';
 import {VersionsManifest} from '../updaters/java/versions-manifest';
 import {Readme} from '../updaters/java/readme';
-import {Version} from './java/version';
-import {fromSemverReleaseType} from './java/bump_type';
 import {JavaUpdate} from '../updaters/java/java_update';
-import {isStableArtifact} from './java/stability';
 import {BranchName} from '../util/branch-name';
 import {PullRequestTitle} from '../util/pull-request-title';
+import {JavaBase} from './java-base';
 
-const CHANGELOG_SECTIONS = [
-  {type: 'feat', section: 'Features'},
-  {type: 'fix', section: 'Bug Fixes'},
-  {type: 'perf', section: 'Performance Improvements'},
-  {type: 'deps', section: 'Dependencies'},
-  {type: 'revert', section: 'Reverts'},
-  {type: 'docs', section: 'Documentation'},
-  {type: 'style', section: 'Styles', hidden: true},
-  {type: 'chore', section: 'Miscellaneous Chores', hidden: true},
-  {type: 'refactor', section: 'Code Refactoring', hidden: true},
-  {type: 'test', section: 'Tests', hidden: true},
-  {type: 'build', section: 'Build System', hidden: true},
-  {type: 'ci', section: 'Continuous Integration', hidden: true},
-];
-
-export class JavaYoshi extends ReleasePR {
-  protected async _run(): Promise<number | undefined> {
-    const versionsManifestContent = await this.gh.getFileContents(
-      'versions.txt'
-    );
-    const currentVersions = VersionsManifest.parseVersions(
-      versionsManifestContent.parsedContent
-    );
-
-    const snapshotNeeded = VersionsManifest.needsSnapshot(
-      versionsManifestContent.parsedContent
-    );
-    if (!this.snapshot) {
-      // if a snapshot is not explicitly requested, decided what type
-      // of release based on whether a snapshot is needed or not
-      this.snapshot = snapshotNeeded;
-    } else if (!snapshotNeeded) {
-      checkpoint(
-        'release asked for a snapshot, but no snapshot is needed',
-        CheckpointType.Failure
-      );
-      return undefined;
-    }
-
-    if (this.snapshot) {
-      this.labels = ['type: process'];
-    }
-
-    const latestTag: GitHubTag | undefined = await this.latestTag();
-    const commits: Commit[] = this.snapshot
-      ? [
-          {
-            sha: 'abc123',
-            message: 'fix: ',
-            files: [],
-          },
-        ]
-      : await this.commits({
-          sha: latestTag ? latestTag.sha : undefined,
-          labels: true,
-        });
-    if (commits.length === 0) {
-      checkpoint(
-        `no commits found since ${
-          latestTag ? latestTag.sha : 'beginning of time'
-        }`,
-        CheckpointType.Failure
-      );
-      return undefined;
-    }
-    let prSHA = commits[0].sha;
-    // Snapshots populate a fake "fix:"" commit, so that they will always
-    // result in a patch update. We still need to know the HEAD sba, so that
-    // we can use this as a starting point for the snapshot PR:
-    if (this.snapshot && latestTag?.sha) {
-      const latestCommit = (
-        await this.commits({
-          sha: latestTag.sha,
-          perPage: 1,
-          labels: true,
-        })
-      )[0];
-      prSHA = latestCommit ? latestCommit.sha : latestTag.sha;
-    }
-
-    const cc = new ConventionalCommits({
-      commits,
-      owner: this.gh.owner,
-      repository: this.gh.repo,
-      bumpMinorPreMajor: this.bumpMinorPreMajor,
-      changelogSections: CHANGELOG_SECTIONS,
-    });
-    const candidate: ReleaseCandidate = await this.coerceReleaseCandidate(
-      cc,
-      latestTag
-    );
-    const candidateVersions = await this.coerceVersions(
-      cc,
-      currentVersions,
-      candidate
-    );
-    let changelogEntry: string = await cc.generateChangelogEntry({
-      version: candidate.version,
-      currentTag: `v${candidate.version}`,
-      previousTag: candidate.previousTag,
-    });
-
-    // snapshot entries are special:
-    // 1. they don't update the README or CHANGELOG.
-    // 2. they always update a patch with the -SNAPSHOT suffix.
-    // 3. they're haunted.
-    if (this.snapshot) {
-      candidate.version = `${candidate.version}-SNAPSHOT`;
-      changelogEntry =
-        '### Updating meta-information for bleeding-edge SNAPSHOT release.';
-    }
-
-    // don't create a release candidate until user facing changes
-    // (fix, feat, BREAKING CHANGE) have been made; a CHANGELOG that's
-    // one line is a good indicator that there were no interesting commits.
-    if (this.changelogEmpty(changelogEntry) && !this.snapshot) {
-      checkpoint(
-        `no user facing commits found since ${
-          latestTag ? latestTag.sha : 'beginning of time'
-        }`,
-        CheckpointType.Failure
-      );
-      return undefined;
-    }
-
+export class JavaYoshi extends JavaBase {
+  protected async buildUpdates(
+    changelogEntry: string,
+    candidateVersions: VersionsMap,
+    candidate: ReleaseCandidate,
+    packageName: PackageName
+  ): Promise<Update[]> {
     const updates: Update[] = [];
-
-    const packageName = await this.getPackageName();
     if (!this.snapshot) {
       updates.push(
         new Changelog({
@@ -192,7 +66,6 @@ export class JavaYoshi extends ReleasePR {
           versions: candidateVersions,
           version: candidate.version,
           packageName: packageName.name,
-          contents: versionsManifestContent,
         })
       );
     }
@@ -204,7 +77,7 @@ export class JavaYoshi extends ReleasePR {
         versions: candidateVersions,
         version: candidate.version,
         packageName: packageName.name,
-        contents: versionsManifestContent,
+        contents: await this.getVersionManifestContent(),
       })
     );
 
@@ -252,43 +125,7 @@ export class JavaYoshi extends ReleasePR {
         })
       );
     });
-
-    return await this.openPR({
-      sha: prSHA!,
-      changelogEntry: `${changelogEntry}\n---\n`,
-      updates,
-      version: candidate.version,
-      includePackageName: this.monorepoTags,
-    });
-  }
-
-  protected supportsSnapshots(): boolean {
-    return true;
-  }
-
-  defaultInitialVersion(): string {
-    return '0.1.0';
-  }
-
-  protected async coerceVersions(
-    cc: ConventionalCommits,
-    currentVersions: VersionsMap,
-    candidate: ReleaseCandidate
-  ): Promise<VersionsMap> {
-    const newVersions: VersionsMap = new Map<string, string>();
-    for (const [k, version] of currentVersions) {
-      if (candidate.version === '1.0.0' && isStableArtifact(k)) {
-        newVersions.set(k, '1.0.0');
-      } else {
-        const bump = await cc.suggestBump(version);
-        const newVersion = Version.parse(version);
-        newVersion.bump(
-          this.snapshot ? 'snapshot' : fromSemverReleaseType(bump.releaseType)
-        );
-        newVersions.set(k, newVersion.toString());
-      }
-    }
-    return newVersions;
+    return updates;
   }
 
   // Begin release configuration
